app/backend/requirements.txt
flask
flask-cors
pymysql
werkzeug
python-dotenv
python-jose